requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
django-cors-headers==4.3.0
django-filter==23.5
python-decouple==3.8
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
jdatetime==4.1.0
django-jalali==6.0.0
elasticsearch-dsl==8.18.0
django-elasticsearch-dsl==8.2
# Object Storage (Arvan Cloud S3-compatible)
django-storages==1.14.2
boto3==1.34.18
# Production server
gunicorn==21.2.0
# Environment variables
python-decouple==3.8
# Database client for MySQL
pymysql==1.1.0

